Pull out the getParser method into it's own file.
cli.js is getting complex, this is a simple bit that can go solo.
